requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0

# Configuration and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# Tokenization
tiktoken>=0.5.0

# Optional dependencies for pretrained models
transformers>=4.30.0

# Visualization
matplotlib>=3.7.0

# Monitoring (optional)
wandb>=0.15.0

# Development tools
requests>=2.31.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

